refactor: reorganize dashboard and layout
