+++ estoque-vendas/src/components/Charts/index.js
@@ -1,0 +1,3 @@
+export default function DashboardCharts() {
+  return null;
+}
